Add some indexes that are missing.
These are on relatively small tables so haven't lept out but I hit a
query I couldn't due due to the missing index on
civicrm_mailing_stats.mailing_id and figured while I'm at it I should
add to civicrm_group. The table is small so I only noticed the lack of
index when the server was on go-slow. Both tables are small enough this
should be near-instant.

Bug: T236955
Change-Id: I0486f13ecd3e8c04538a2d3afbe4ce4bd864ffee