Report recurring stats to grafana via hook
Needs smoke testing.

This is a quick way to get counts of completed and failed txns after
each run of the recurring charge job. Another way to do it would be
to run the hook after each charge and send much more data through it,
so the function that attaches to the hook could decide how it wants
to aggregate (e.g. by payment processor id, by payment instrument).

Bug: T199390
Change-Id: I3e95d7aa9efcf50837645332f8092d0d8e6e2b89