Switch to getting the name of the temp tables by function.
The intent is to switch over to using the CRM_Utils_SQL_TempTable class to generate unique temp table names. This function
/ class has a couple of helpers including ensuring utf8 consistency & autodrop as well as a nice syntax. In addition
it gives the unique temp table names a consistent prefix. This is not hugely important now but as we switch to having a
read only DB connection it gives us the opportunity to filter them from replication.

In order to make this change we need to switch from using the drupal query function to the Civi function. This is mostly
good as it is in line with our desire to migrate Civi code into Civi extensions and the db_switcher code is pretty hacky.
This doesn't do that - it just helps make the commit that does a bit smaller

Bug: T238691
Change-Id: I1540ac80d64746a63254c69ac3c1424bf8b7108f