Switch to unique temporary table names
We need unique temporary table names per process of we risk having a donor services person trigger an email & it flushes an in-progress eoy_receipt job.
This uses a Civi helper class to generate them according to a naming convention designed to allow us to exclude temporary & temporary-like (functionally
temporary but not at the mysql level) tables from replication. This counts as best practice in the civi env I think

Bug: T238691

Change-Id: I55b550aea4e303031a1d7581b99c3e4c879918bc