Throw database contention exception on constraint errors
In places they are known to be deadlock caused

When an email or address or contribution fails to insert with a
constraint violation the only known cause is a deadlock + rollback.
This changes the error code so they hit the retry queue

I found it was possible to replicate & breakpoint these in phpstorm.
You can run 2 lots of unit tests at once in phpstorm (not sure if
config is required - for me it's just right-click but they have to be
different). I've been running BenevityTest in one tab & the full suite
in a second tab & been hitting break points well short of
death-by-boredom

Bug: T182148
Change-Id: Ic4aa561984c753d39861302b49c8d40b7d4b8d5f