Only show cancel instructions when donation is active
Many of the donors receiving this email have already canceled their
donations, and MBeat thinks that including this line could be alarming
and imply that their donation is still active. Also, for the on-demand
button that uses the same template, there's no guarantee that the donor
ever had a recurring donation.

Change-Id: I98bc486f922def6924bddbae1c3cb41333a18a95