Fix bug where stretches of time were being skipped
We were hitting a bug specifically when there was a large file
that took multiple crons to process it AND we were caught up.
Despite having a 'progress_end_timestamp' to track where the file
was up to we were re-calculating the end date. This is fine when
we were not up-to-date because it was correctly calculating it.

However, when we are up to date the calculate is the lesser of
the calculated value & now. The issue occurred when now was
greater than the progress_end_timestamp - ie. later than that the
value of now when the file was first requested.

If the entire file was processed in one hit all was fine but if
it hit a limit & then a second job continued on the same file it
was recalculating the less of the calculated value & now & coming
up with a too-high value for 'now' compared to the value when
requesting the file.

Bug: T236855
Change-Id: I0aa65f7062344978a2540d71b3e6a18157bc8733