Add basic thank-you extension
This  exposes our thank you logic as a civi api and creates a very simple form for it which is accessible
as an action link from the contribution table (ie. on a contact's contribution tab or search results, on the right).

I didn't attempt to add it to the form because I think as long as it's accessible from one place.

I feel like really we should have this in the same extension as the similar one Jack did - but the tile felt
wrong.

Bug: T227903
Change-Id: Ia0a9565aeca88d890d29429ddf92fb26e0cfccc4