Update contribution recur statuses.
Use the new statuses we negotiated in Barcelona so that we
are no longer misusing the old ones (& nno longer need civi hacks to support that).

Note this is likely to also address Michael's issues about not being able to cancel
'Failed' because they will be 'Failing' instead - which is not caught.

Bug: T244326

Change-Id: Ie18bbbcb77a20780581adfee953382aefe7a10f7