TY test hack to render pt_PT correctly
Our files use 'pt' to mean Portuguese as written in Portugal and
'pt-br' to mean Portuguese as written in Brazil. CLDR uses 'pt' to
mean Brazilian Portuguese and 'pt_PT' to mean the one in Portugal.

CLDR has different currency formatting for the two. This will let
us test the thank_you.pt.html with the currency formatting from
Portugal.

Change-Id: Iba851983557ea7e879b32e067bb9cb43d68d0238