Add new ckeditor5 extension
These last CiviCRM hacks have lasted this long for a reason - they are hard. In the case of our civicrm.settings.location.php
it was to support kcfinder which is the file manager in CKEditor. kcfinder is brittle & was last edited in 2014 but I found elfinder
as an alternative, which is integrated as an option here. However, for our purposes the base64-uploader (which I've labelled embedded images)
and which is available as part of a ckeditor5 build is actually a better option. Last time people wanted to add a signature I had to help
them to embed an image as external users can't see images on our Civi server. So while the elFinder option will hopefully improve
community uptake the embedded option is the one for us.

Note I have not added to the install script since I think we'll get conflicts with other things in review.

Bug: T244323
Change-Id: Ib3afcc930646c247716bece0ed1ffd728474c9ac