Adapt SmashPig recurring processor to PaymentProcessor interface
Updates calls to createPayment and to approvePayment to pass
normalized parameters and to deal with normalized
CreatePaymentResponse and ApprovePaymentResponse objects.

Improves error handling - now when an error code deemed non-
retryable is recieved, we immediately cancel the recurring
contribution.

Also updates the namespace of a moved SmashPig enum and uses a
more appropriate Exception subclass for the charging-too-soon case.

Bug: T242278
Change-Id: I4e2e7ad97d27ccba9a43769d02d3c37e3489303d