Consolidate email location handling & use core api to add emails
We have not been passing in email to contact.create. This is a problem on new
inserts where we have no name fields & hence it bounces. The rationale for
separately adding the email, rather than as part of the contact.create
call is historical & the only thing it really does that the other doesn't is ensure is_billing = 1.

However, is_billing is not used much in core or our code. It feels like 'a good
idea at the time' & in theory can be used to cause a non-primary address to be
preferred for some purposes. I don't think it happens much in practice & really
 it only matters when billing is not the same as primary

Bug: T246889

Change-Id: I151e7f9288f18a7a4201b3fbed800fa5c089736d