Updates to SmashPig to support recurring iDEAL
Gets different PaymentProvider classes based on the payment_instrument,
and only uses approvePayment when necessary. With Adyen, the gateway
transaction ID we want to store is the last one that comes back. I.e.
for card transactions, we want the ID from the approvePayment call.
For iDEAL donations, we don't need to make an approvePayment call,
so we store the ID from the createPayment call.

Bug: T230837
Change-Id: Ie0762a7a926cb0de5bef7b78a326bfcd1a235074