Update SmashPig, rename log_id in audits
This update is mostly to get a fix for the Adyen createPayment
status normalization, but a SmashPig patch that was merged previously
necessitates another small change.

In SmashPig change I054c1be200dbf7371b6292 we started setting the
invoice_id field to the merchant reference / our-side order_id.

We need this adjustment on the CRM side to be able to deploy the
change. It should be fine to pass the invoice_id along as part
of the merged data, so we delete the merge_data overrides that only
existed to unset the log_id.

Bug: T250106
Change-Id: I4777a780f7b0bca92eb74e3a6c9e055feb710de6