Fix a couple of issues installing and testing
Trying to install wmf_civicrm standalone (not in the same drush en
list as wmf_common) failed with a CRM_Core_Config not found. Adding
the initialize() call fixed it.

The WmfTransactionTest is failing because it's trying to access an
array offset on an integer (callAPISuccessGetCount, called under the
hood, always returns an int)

Change-Id: Id84f38a45dab010e3508fd51b7364fd5cc4e8cdd