Don't update c_t rows with more than contribution_id
This could indicate a failure in sequence generation. We only expect
to insert the full row and then update it by adding a contribuion_id.

Also, if a contribution tracking row already has a contribution_id,
throw an error if we're trying to change that to point to a different
contribution_id.

Cleans up existing CT queue consumer tests a bit, using new expected
exception syntax and generic ID cleanup.

Bug: T253250
Change-Id: I82e3fcbbb4bdb3345339166a16e0417439e46ba3