Update composer package for silverpop retrieval
This adds the getQuery action to the underlying silverpop repo
+ a few travis fixes.

https://github.com/mrmarkfrench/silverpop-php-connector/pull/31

This change only adds an action - it can be deployed safely with no change
occurring. There is a unit test merged to the repo ^^ for the
new action, which is really just the straightforward part.

Making something sensible of it is for the omnimail/silverpop
repo and then the extension gets it to Civi

Bug: T252247

Change-Id: I2d61830dac0ab53ba910f11d66f31f4f513bd0e2