Add second field to store api retrieved criteria
The format it comes back in is a bit funky & I think we might have to work through some
parent queries and weird things. So my thinking is to get and store both and monitor for a while.

We can turn off Omnihell once we are confident there is parity.

Note the command to upgrade is

drush cvapi Extension.upgrade

& then

drush cvapi Omnimailing.load
or
drush cvapi Omnimailing.get
(to not put it in the DB)

Bug: T252247
Change-Id: I47e96cde60f84daf04412735a805aec578c3df29