Send message from failed recurring if smashpig_recurring_send_failure_email set to true.
This is the basic functional code to send failure emails for recurring contributions that fails. It requires
the template files added to be added to the message_template table as per
https://gerrit.wikimedia.org/r/598908.

Note that without the template this can be installed with no change. With the template present it
is hack-hobbled to only do English language. Two apis allow the Email to be previewed / sent:

FailureMessage.render
FailureMessage.send

While I haven't yet tested this should be accessible in js using CRM.api4.
See tests for See tests for example php usage.

https://docs.civicrm.org/dev/en/latest/api/interfaces/

See tests for example usage.

Areas that need follow up /consideration
1) Maggie is working on a preview button
2) The currency formatting is a bit of a stretch right now - it calls our Twig class
3) I have reservations around the token names - we have 2 custom types.
  a) {now.MMMM} - this formats now in a localised way using format http://userguide.icu-project.org/formatparse/datetime
     I have misgivings about the icu format as elsewhere in Civi we use strotime formats. I think it's OK for
     now but I'm holding open the option of changing it later
  b) {contributionRecur.amount__formatted_money}. It seems like the regex wont't allow other symbols so I'm using
     the double underscore. Not sure if there is a better idea for that either.
5) The code so it will only send in English for now - pending provision of other languages
6) The text talks about unsubscribe but not how - I guess I need something like
   Wikimedia Foundation, you can <a href="{{ unsubscribe_link | raw }}">unsubscribe instantly</a>
  which is not coded. Michael said just telling them to use return mail may be best
8) Note this is being done from Smash Pig - we could alternatively do it from a post Hook when a contribution_recur
is updated to Failed. I'm not sure if there are risks of accidental triggering then

Bug: T230064

Change-Id: I25f071843621f7bd655612de8db6b4df0364c43c