Remove submit confirmation
This was breaking the form buttons under the shoreditch screen - an
extraneous quote after  'want to send this email' was causing validation /
js errors.

However, after I fixed it the submit block wasn't popping up.

I feel like we can remove this altogether as it probably is a bit
towards over-asking, since they have already clicked 'send' and I also
recall this as a tricky thing to get to work.

I added a cancel button so it's clearer from the form how to back out

Bug: T230064
Change-Id: Ife02d997f98939361e58d75057d4688ecbfc398b