Filter matching gift policies on type
As described in an email, we can filter by categories using the
category name as a QS variable with value 'yes'. Since it looks
like there's no way to do an 'OR' search for multiple categories,
we instead do multiple searches and merge the results.

Add this to your settings to test:

$civicrm_setting['domain']
  ['matchinggifts.ssbinfo_matched_categories'] = [
    'educational_services',
    'educational_funds',
    'libraries',
  ];

Also, any credentials you may have had in the nested setting
['Matching Gifts Data Preferences']['ssbinfo_credentials'] should
be moved to setting
['domain']['matchinggifts.ssbinfo_credentials']

Bug: T249924
Change-Id: I63214ba9d84e609785883cb9f16c92c9ea3baf03