Move MergeTest to new wmf-civicrm
As part of consolidating this code into the extension I have moved the test. There was a test
I felt didn't really belong with the MergeTest so I renamed the remaining part BlankAddressTest
(it also had some complex dependencies which made me not want to move it to the
extension) and there were 2 helper functions on the BaseTest class. One was only called from MergeTest
so I moved that and the other I rewrote more simply (using apiv4 for the test class).

I also added a check to mmake sure contacts are deleted in the tear down.

Some of our tests and code are now in the deduper but they test deduper functionality. The merge test
tests our requirements - which is a combination of the deduper doing what it says on the box
and the right settings being configured - so there is a place for both.

I want to add new tests for email handling but I want things to be 'in the right place' first

Bug: T256316

Change-Id: I4386b06d5406118f4f6d9a50693be63135f12e85