Move email handling to deduper, fix email loss bug
The code here should facilitate email, phone and address. However, if it is
only tested and enabled for email.  The default i the Dedupe.setting of 1 for
the deduper_resolver_email is preferred_contact_with_re-assign and once caches
are cleared it should be activated. I think this would a reasonable default for
all sites. I think for phone there is complexity around phone type & for
addresses there is more complexity so I would leave the default for them as
null & formally enable for us.

The second part of enabling is to add bypasses in the code in wmf_civicrm which
is the handling we are moving away from. At the time it was written the
wmf_civicrm handling seemed like the best we could do with the state of the
code at the time but we continue to see intermittant issues with contacts
having no primary location type. The deduper uses a different methodology -
altering the calclulated information that the dedupe process has. I believe
this was the intent of how the hooks would be used in the first instance but a
lot of what is happening is us working around nasty core handling (but actually
it used to be much worse).

In terms of the specific bug this is intended to solve - I think it's
reasonably explained in the code comments for testMergeEmailNonPrimary and
running that test with breakpoints in
CRM_Deduper_BAO_Resolver_PreferredContactLocationResolver::resolveConflicts is
probably the best way to review it.  There is also coverage of a range of
combinations in testBatchMergesAddressesHook

Next steps would be to try to extend the test to cover address and phone and
fully remove the code from wmf_civicrm.module and enable the deduper code to be
used for them.

Bug: T256316
Change-Id: Ic9a6163de7399e9f53864e9c758e3a62a273bfbe