Remove blank address test
The premise of this test is 'dedupe will overwrite good addresses with blank ones & we have a
command to fix that if needed' - lets's test the command.

However, once we fix it so dedupe won't overwrite good addresses with blank ones we
can no longer create the bad data to test. The fix for that is in
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/618856/12
but I have separated this out as something in it seems to matter for later tests
(probably the logging enabling works here & maybe not in the parent)

Change-Id: I155f9b0b1359bf566816bcb06d9d1176094ad026