Add config to support rpow in our dev environments.
In conjunction with https://github.com/civicrm/civicrm-buildkit/pull/543
this ensures that logging is set up on dev environments and that
we can cope with RPOW for it.

In dev environments files in civicrm.settings.d are loaded when loading civicrm.settings.

This does not happen on prod/staging so it's an OK place to put things we want to add in - in
this case handling for the logging DSN.

The buildkit PR allows us to build up a json file of settings to load in dev environments on install.
These are not override settings but install settings (ie we expect they could be altered in the
course of development / unit tests.

Bug: T237350
Change-Id: I86e338a6ff7b99366f8eca1517ea54f903f384bf