Fix SmashpigTest installments to 0 (unlimited) in test
Our test set up was setting installments to 1 but expecting that after the first payment
the next_scheduled_date would be extended. In fact the date is now rightly not extended in core
as after 1 payment no more are expected. This updates our test set up to match what we are testing
(which otherwise fails on 5.29)

Bug: T259578

Change-Id: I13365b8a90d22c5f595d68c3eb8d58e47b19f559