Add 'pretend' FK to MailingProviderData so entities will be updated on merge
Note the change in the xml is the 'real' change and then the DAO is regenerated with that change.

We don't want a real FK just for volume reasons but a pretend one means Civi understands that the contact_id field is the contact.id
at the php level.

A real live one would also exist in the sql table

Bug: T220149

Change-Id: Iffc8b2cf3ae914528a93a9717dd6397cb45963fd