Remove code that adds indexes to civicrm_entity_tag
These lines Do make the indexes closer to, but not the same as, production. However the underlying reason is gone.

On production the unique index is

  UNIQUE KEY  (,,),

but on new installs

  UNIQUE KEY  (,,),

The original sin is that the early version had entity_table first meaning that entity_id
was only index-accessible if both were provided (and the less unique one was first).

Given this doesn't solve a dev indexing issue or make dev environments the same as
production I think this no longer achieves anything

Bug: T261359
Change-Id: Ib152bae0ef319827a35179c8d5143e5b54d7da6b