Remove statement to alter preferred_language length from upgrade
The immediate reason is tests seem to be failing on the db_switcher and we already have intent to remove it.

However, I'm making the case to remove this rather than updated it because I think it's actually good if our locals
are a bit stricter than prod here. The reason seems to be that we have some language variants longer than the expected 5
chars. However, there are 30k records in total and I think mostly older. If there is an issue here that comes up again
I think we want to figure out what the right answer is (ie. core is too strict or we are doing something weird) and
having less tolerance on dev environments plays into that logic

select DISTINCT preferred_language,  length(preferred_language) FROM civicrm_contact WHERE length(preferred_language) > 5;
+--------------------+----------------------------+
| preferred_language | length(preferred_language) |
+--------------------+----------------------------+
| ang_HK             |                          6 |
| be_taras           |                          8 |
| enp_US             |                          6 |
| es_419             |                          6 |
| fiu_vro            |                          7 |
| hak_CN             |                          6 |
| nds_nl             |                          6 |
| pcd_FR             |                          6 |
| yue_CN             |                          6 |
| zh_hans            |                          7 |
| zh_hant            |                          7 |

Bug: T261359

Change-Id: I92eaa211fd37b2ea8a62807f4d19cf0f1128b3b7