Remove upgrade commands for wmf_fredge_qc
The schema is accurately added on install due to 7001 calling the definition. The civi upgrade
(likely because of rpow interacting with the db packages change) is hating wmf_civicrm_get_dbs()

In these 2 functions it seems simply redundant & historical

Note there is a slight discrepancy between the update command and the index on payments_fraud_breakdown.

I confirmed the schema version reflects live

Bug: T261359

Change-Id: Ie928b46c7095dd245764e414c3792208ceeaa36e