Join the brave new world of 2010
Per https://api.drupal.org/comment/11769#comment-11769 it is no longer necessary to do more than
declare the hook_schema to have tables added on install. Somewhat before that it
also became possible to define prefixes for tables. I have added these to
buildkit here https://github.com/civicrm/civicrm-buildkit/pull/556 - it will be necessary
to add similar lines to prod civicrm.settings.php (but not in the critical path on this
commit as we have already installed the extension on live

Change-Id: I00ef90635637b6029e75a73f2081bffaeeb80cea