Insert paypal recur records on first payments
There are a whole lot of MISSING_PREDECESSOR messages in the damaged
table right now. It looks like PayPal is just not sending us the
subscr_signup messages we expect to get before the first payment in
a series. This patch will automatically add the contribution_recur
row when it gets the first payment in a series, since the payment
message contains everything we need except the frequency. We assume
that to be monthly, as that's the only recurring frequency we offer.

Bug: T248420
Change-Id: Icd7d1c73ad7b6a34f1b1fae2e1382310e7606d4a