Start the process of moving our mailer class to Omnimail
I was reluctant to do any work building this out while it was part of a drupal extension that we want to decommission
so starting the process of migrating to a civi extension.

I removed the drupal mailer as it seems to be clearly unused

I considered a couple of 'homes'
1) phpmailer as a core extension - I think it makes sense to replace the existing Pear
classes in core. However, in terms of agreeing something & planning for that the timeframe
is wrong, not least because Tim is currently scoping out Oauth & credential related work
that would likely impact
2) generic phpmailer or mulltimailer extension that we can publish. To a certain extent
omnimail is / was meant to be that but it has started to look a bit wmf-fy over time. Without resolving
the broader plan for core I felt like starting yet another didn't make sense

3) omnimail extension - this already has phpmailer as part of it's vendor selection. We may or may
not ultimately use the Omnimail libraries and that ambiguity could argue against it but
I felt in the end that omnimail is already pretty wmf-fy so we can use it to migrate
out of our drupal module

Bug: T264663
Change-Id: Ib5f43f5f8a9ae3a915affb1a4b6e772c1fac9ff8