Add api to check Job progress
The intent is that we can add a scheduled job that calls this hourly & if
there is an old entry an exception will be thrown, triggering a fail mail.

I also added the Mailing provider data v4 api while I was at it (we already have v3). This is
exposed in Search Kit & apiv4 explorer

Bug: T222287

Change-Id: I836364d45f57b223bb57faa38e14d631a33d17f0