Make it easy to run tests on the silverpop connector package we use
I tried adding this to our test run but it failed in a way which I don't think is worth the time to
resolve. After all the tests ran through it finally released the classes. The destructor
on the silverpop class then tried to log out and that failed. I was always on the fence
about whether we should include in our tests & this cinches it for me. On the other hand

1) I actually have travis on my repo and
2) This change at least makes it so we can run the tests locally
Change-Id: I3101de59f8f2e43bd279e933915a439b12c400ed

Change-Id: I03dd8606fdcd87fb1e04db1ed340c226489f21b8