Fix install of wmf reports such that they are managed by .mgd files
The modern way to manage entities that are added by extensions is to declare them in
mgd files (see https://docs.civicrm.org/dev/en/latest/hooks/hook_civicrm_managed/).

Bug: T270684
These reports precede that. One of the reports is likely unused (there is an option query to
MG on that) but the other one is still used so it makes sense to deal with both at this stage.

This commit adds the mgd file. However, since the option values already exist on prod
we can't rely on the mgd file to pick that up - so we are manually adding the entry to
the civicrm_managed table to link the existing report_template option_values to the
wmf-civicrm module (even though these reports are not there yet the install portion of
the code can be)

Change-Id: I40921867a067e0c0de1c6d08cc89f3eccab49032