Move gateway reconcilliation report into wmf-civicrm
This report is the only thing in use in the extension I believe. At this stage the wmf lybunt
report is still in the old module and I strongly believe only the core lybunt is used - if that
is confirmed the remainder of the extension can be removed.

The tests on the wmf report module were not running and it turned out they were failing on an
e-notice on the other report so I simply brought over the changes specific to this report

Bug: T270684

Change-Id: Ie5a9c6d4dd0972c61b3c6562ad4a812e65e39725