Move index creating for mailing_stats to mailing_stats extension
This table is created in the extension so it should be the extension that declares the install.

While I was at it I removed references to utf8 in preparation for us moving to
utf8mb4 - for which we should not have collation declared & cleaned up an index.
The fields altered are pretty immaterial to how we use this table

Bug: T270680

Change-Id: Ia96c6f8865c90ecab5feb87b20af68fa19fd7184