Exclude Japanese language emails from eoy_receipts run
The intention of this patch is to skip all Japanese EOY emails & update their status to
deferred_ja - the logic is that this gives us a little more time to check out
whether we are OK with the use of the first_name on the basis it 'probably'
contains the last_name - see phab for more.

By updating the status to 'something other than queued' it will not be processed in
subsequent runs. But we can revert this patch later & db update them
all back to queued and the Japanese ones will go through. It might also be
that the list of queued ids will be helpful in terms of someone eyeballing it

Bug: T271189
Change-Id: I0ece94d2e998ee250b768bff7415a34f8a89f4b6