Remove fixtures use
The fixtures idea was probably really clever. But the fixtures are created
before any other tests run and are only torn down after every test is run
when they are created in the message provider.

In fact it would be fairer to say 'there is an attempt to remove them
after the test has run' as it would rely on the
test class being destroyed so that the fixtures
destructor is called. Not reliable

Bug: T274943

Change-Id: Ic913a6715d25c4f86f848609ffb8eaabf5625fcd