Update Afform & dependencies
There was a major update to afform in the last release which was
a breaking change on contactlayouteditor & deduper. Oddly they
didn't break. I now know it's because we had a version of afform in
our extensions directory & were not using the core one

If pulling in this patch a bit of cache clearing is in order. I
*think* drush @wmff cvapi System.flush is enough - I've also
been hittin 'refresh' on the extensions page.

This also includes some bug fixes that I did while testing - the
commits are
https://github.com/eileenmcnaughton/deduper/compare/5163f74527686...master

I believe that one of the bugs I was hitting was
Bug: T275668 - it was a tricky issue to diagnose where the code to
remove a deleted-by-merge contact from the listing didn't work
if there were 2 rows in a row to remove - due to the way the array was being
iterated. It seems that in the jquery each if you remove item
2 from an array using splice then item 3 becomes 2 and
item 3 is effectively skipped in the loop
https://github.com/eileenmcnaughton/deduper/commit/f3d608256c8abe356bd423a4dca76832e5236ce0
seems to fix

I also
1) fixed T275663 (which was pretty trivial)
2) added street_address & city to the display if present (read only) -because
I was worried DS might be merging matches based on meaningless data (eg
clearly made up addresses) without this visibility.
3) other bits & pieces I hit in testing

Note that the above list may need a little massage to
show on live when we deploy as we have overridden the afform
on live (there is an afform edit mechanism) to add the partner
field.

Bug: T275728
Change-Id: I52f3968f8bd79678b7c4b0cc29f45dc05719330d