Remove now-obsolete code
This was added as part of https://phabricator.wikimedia.org/T116420
when the UI changes in 4.6 confused users. In the years since then the field
was renamed & obviously no one noticed when it stopped working.

I'm trying to clean up this function in order to move it to the
extension and make a change so we can remove our last civi-hack when
upgrading civi

Bug: T277590
Change-Id: I13bd016b3d4294e5f16e15332c71befc47f2e5f7