Remove 'Contributions OR Soft Credits?' from search
Remove the field 'Contributions OR Soft Credits?' from the contribution search
and advanced search pages.

When users attempt to search on soft credits CiviCRM creates a humungous temporary table.
This kills the server. We have hacked core to throw an exception but this
moves the hack to the extension and makes it less taunting to the users.

Removes it from both advanced search and contribution search.

Preparation for
Bug: T277590

Change-Id: I4cd37416f2e4fa15ed103026f6c23eb53c1c9556