Remove WMF_lybunt report
Once we have deployed this we can delete the actual code (this should
cleanup the references to it).

We already asked the question as to whether is is still used. We have 2 answers

1) It doesn't seem to load due to lack of query optimisation. I did optimisation
in the core report a while back and I think since then we have only used
the core report
2) MG says it would only be used it it were a dashlet https://phabricator.wikimedia.org/T270684#6713543
In order for that to be the case it would need a row in civicrm_report_instance - which it does not

select id,report_id FROM civicrm.civicrm_report_instance WHERE report_id LIKE '%wmf%'

Bug: T270684
Change-Id: I0a74c04475e36398d46c16fea7ceef902bba6a97