Add test and fix for handling emails
This addresses the issue where our behaviour for multiple emails
is not working.

We expect that the resulting contact will have a primary email
that is consistent with that of the preferred contact.

The email from the other contact will be moved over but will not
be primary

Note this alters existing tests as we will be keeping more emails
now

It's hard to put into words how this works but ultimately the
BAO_Merger class considers whether
migrationInfo['move_location_email_0'] is set and if so it considers
the values in migrationInfo['location_blocks'] as form instructions
which might include the eventual location type id, whether to
overwrite the prexisting one (by virtue of whether is has a
'mainContactId' key combined with the 'override' value) and which
one to set as primary - 'set_other_primary' being the main key here

Bug: T276391

Change-Id: I06f472c6a5d37ac401ef0c0a96e2bb66ee5bdc0f