Move rest of dedupe handling out of wmf_civicrm.module
This last bit of code is still here because it is confusing. However, we do have good test cover
so I've relied pretty heavily on testing + memory. My recollection is that we started with fairly
complex location address merge rules but over time they were reduced to basically keeping
that of the latest donor

Bug: T270679
Change-Id: I22a64b0128590fac4e360f80c27da11486dbb705