Delete unused fn, log to see if we can delete another
wmf_get_calculated_field_names is newly unused after Eileen's
recent cleanup.

The post contribution hook is calculating some things that look like
they are all now handled in the wmf_donor trigger. Happily, it's
already checking to see if the existing data matches its calculations
so we can just log when it doesn't to see if it's ever necessary.

Change-Id: Ia4c6244782710423b5fad05e39ef7f201691ecd7