Add monolog extension
Note that I'm just adding this at this stage - but when we enable
we need to configure the right monolog loggers - see the readme for what we get by default

To add a wmf logger we will need to add a new monolog logger and then
we can, for example, do Civi::log('wmf')->debug('merging contacts 1 & 2')

And that will be output to whatever is configured for channel 'wmf'
with debug or lower as minimum severity.

The firephp thing is kinda fun.

Known issues
- I planned to replace failmail too but we have an out of date version of phpmailer so I didn't
bring it into scope on this round
- Error on disable https://github.com/civicrm/civicrm-core/pull/20144
- Notice on enable https://github.com/civicrm/civicrm-core/pull/20122
- I wanted to create a full edit form but pending https://lab.civicrm.org/dev/core/-/issues/2567
- And I wanted to display the configured below the add - https://lab.civicrm.org/dev/core/-/issues/2569

Current commit https://lab.civicrm.org/extensions/monolog/-/commit/d7ff4a06c28af7cb84007e992b471566b995df67


Bug: T279983

Change-Id: I981513ea506a6484871b9db13d769bbf8c7927b6