Keep imported org details on update
Prior to this patch custom values for imported organizations were being added to the organization
when newly created but not when updating

This changes the update behaviour to append the new value to the existing value stored for the contact,
unless the new value is already a substring of the existing one if the donation is over 25k.
This might create slightly messier data
(e.g if the new value is variant of the old one) but all data will be preserved
and for donations over 25k we can expect manual review. For those under we just overwrite.
Bug: T278892

Change-Id: I44189bc386bdcbb49aa01458821b9abc972a5566