Remove more unused code
With this change I've moved the phab into review as I think this is 'done enough'
for the context of this ticket. We culled out the low hanging fruit. There is still
more to remove but some requires a bit more thought & leaving to next round seems fine to me

Bug: T270678
Change-Id: I52cb7bb06760d7f679fd57c4064cbcd4b73bba7b