Integrity fixes to the search & lock file
Note the lock file is not used by us but is packaged in the extension

This search clean up fixes the issue of the monolog search display not editing
correctly. I might need to figure out some sort of manual update but
if you already have monolog installed locally then this won't pull in
unless you delete the saved search row from civicrm_saved_search and
civicrm_search_display and civicrm_managed (where entity = SavedSearch).

You then need to get them to regenerate the managed entities
This seems to happen when doing a menu rebuild in the UI
but not during a system.flush

Change-Id: I930eda49381f40908612c9cbf5710d54af1dd38b