Move alterLogTables & alterTriggerInfo to the extension
This is just preliminary cleanup before working on a new trigger
for the benefactor field. I'm just moving trigger related
code to the extension.

No actual change in this code but I did add some comments.
Note that if this code is still working then turning
triggers off & on again will not result in triggers
being created for tables like
civicrm_financial_item (or if they are they won't be triggers
to log to the log_civicrm_financial_item table)

For the second function move we should still see triggers created
to calculate the wmf_donor totals
Bug: T281268
Change-Id: I71ac569b0b77aa1bcab327966beaaa91d64fddd4

Change-Id: I911d08f45573d34861c26b99a03a92e3f27e27b0