Update family name pair table
This makes the changes that came from my neighbour & per my email

I also fixed up the MergeConflict table a little as it did
weird stuff when I regenerated the boilerplate.

Note that you need to test the upgrade script on a site with data in the table,
which local dev environments might not have. Staging does, or I can
help you through hackery to get there
Bug: T281493
Change-Id: I52ea8e2d396fb2171639ed428469854ed18c14f9

Change-Id: I0b2ef4da1cf18692a4d421999375418f016de441