Update contactlayout editor
The screen for this extension is at Administer->Customize Data and Screens->Contact Summary Layouts

There is currently a deprecation notice on this screen which can be fixed by updating the
extension. Note that you need to rebuild menus by hitting the url civicrm/menu/rebuild
after pulling the code.

Also run the extension update
(reachable from https://wmff.localhost:32353/civicrm/a/#/status)

ALso included is the relationshipblock extension update
It's mostly just 'noise' but there is a patch in there for the later
contactlayouteditor

https://github.com/eileenmcnaughton/org.wikimedia.relationshipblock/commits/master
Bug: T283559

Change-Id: I3c5a5f025a11c4b633a3f76b09bd28d8f16034ed